Move core files into dedicated folder
