+++ lib/core/bit-buffer.js
