+++ lib/core/byte-data.js
