+++ lib/core/error-correction-level.js
