+++ lib/core/galois-field.js
