+++ lib/core/mask-pattern.js
