+++ lib/core/mode.js
