+++ lib/core/polynomial.js
