+++ lib/core/qrcode.js
