+++ lib/core/reed-solomon.js
