+++ lib/qrcode-draw.js
@@ -16,7 +16,7 @@
 */
 
 var bops = require('bops')
-var QRCodeLib = require('./qrcode.js');
+var QRCodeLib = require('./core/qrcode.js');
 var QRVersionCapacityTable = require('./qrcapacitytable.js').QRCapacityTable;
 var QRCode = QRCodeLib.QRCode;
